Add files via upload
Exercise 4.4: Monte Carlo simulation of transcriptional pausing 

The position an enzyme has on a strand of DNA can be modeled by a random walk. The goal was to write a function that would simulate the number of steps it would take for an enzyme to get from one position to another. 

Reference : https://justinbois.github.io/bootcamp/2019/exercises/e04.html