Add files via upload
Problem 5.1: Growth curves from a movie

One way of estimate the rate of growth of bacteria is calculating the area under a microscope. The goal is to write a script that takes a video file of bacteria that express GFP, calculates the area in the image the corresponds to bacteria, and plot it as a function of time. 
Source: https://justinbois.github.io/bootcamp/2019/exercises/e05.html